fix: adjust create color of mandelbrot

+++ includes/utils.h
@@ -3,7 +3,7 @@
 
 # include "fractol.h"
 
-int	ft_create_trgb(int iter);
+int	ft_create_trgb(double iter);
 void ft_mlx_pixel_put(t_vars *data, int x, int y, int color);
 int ft_render_fractol(t_vars *vars);
 int ft_destroy_all(t_vars *vars);
